v2.0: Finalized Release
This release has breaking changes fully detailed in CHANGELOG.md. Rather than
adding _lots_ of new functions, I have dropped the number of functions that are
included and merged a few of the functions into other functions. The reason is
that I was not finding myself using most of those functions.